add changelog & fix unit test
