+++ changes/en-us/2.x.md
@@ -12,6 +12,7 @@ Add changes here for all PR submitted to the 2.x branch.
 
 - [[#7104](https://github.com/apache/incubator-seata/pull/7104)] fix impl of supportsSourceType is not defined
 - [[#7112](https://github.com/apache/incubator-seata/pull/7112)] bugfix: remove the condition that IPv6 must start with fe80
+- [[#7107](https://github.com/apache/incubator-seata/pull/7107)] fix the issue of failing to parse annotations in TCC mode when the business object is a proxy object.
 
 
 ### optimize:
@@ -47,6 +48,6 @@ Thanks to these contributors for their code commits. Please report an unintended
 - [funky-eyes](https://github.com/funky-eyes)
 - [psxjoy](https://github.com/psxjoy)
 - [xiaoxiangyeyu0](https://github.com/xiaoxiangyeyu0)
-
+- [wxrqforever](https://github.com/wxrqforever)
 
 Also, we receive many valuable issues, questions and advices from our community. Thanks for you all.
