+++ changes/zh-cn/2.x.md
@@ -12,6 +12,7 @@
 
 - [[#7104](https://github.com/apache/incubator-seata/pull/7104)] 修复SeataApplicationListener在低版本springboot未实现supportsSourceType方法的问题
 - [[#7112](https://github.com/apache/incubator-seata/pull/7112)] 校验是否IPv6网络ip取消必须以fe80开始的条件
+- [[#7107](https://github.com/apache/incubator-seata/pull/7107)] 修复tcc模式下，当业务对象为代理对象时，解析注解失败问题。
 
 ### optimize:
 
@@ -46,5 +47,6 @@
 - [funky-eyes](https://github.com/funky-eyes)
 - [psxjoy](https://github.com/psxjoy)
 - [xiaoxiangyeyu0](https://github.com/xiaoxiangyeyu0)
+- [wxrqforever](https://github.com/wxrqforever)
 
 同时，我们收到了社区反馈的很多有价值的issue和建议，非常感谢大家。
